arguflow-server-client/pyproject.toml
[tool.poetry]
name = "arguflow-server-client"
version = "1.0.0"
description = "A client library for accessing arguflow-server"
authors = []
readme = "README.md"
packages = [
    {include = "arguflow_server_client"},
]
include = ["CHANGELOG.md", "arguflow_server_client/py.typed"]


[tool.poetry.dependencies]
python = "^3.8"
httpx = ">=0.20.0,<0.27.0"
attrs = ">=21.3.0"
python-dateutil = "^2.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = ["F", "I", "UP"]
line-length = 120
